+++ tests/testthat/test_classif_xgboost.R
@@ -12,3 +12,17 @@ test_that("xgboost with softmax", {
   result = run_autotest(learner, predict_types = "response")
   expect_true(result, info = result$error)
 })
+
+test_that("xgboost with multi:softprob", {
+  task = mlr_tasks$get("sonar")
+  learner = mlr3::lrn("classif.xgboost", nrounds = 5L, objective = "multi:softprob")
+  p = learner$train(task)$predict(task)
+  expect_equal(unname(p$score()), 0)
+})
+
+test_that("xgboost with binary:logistic", {
+  task = mlr_tasks$get("sonar")
+  learner = mlr3::lrn("classif.xgboost", nrounds = 5L)
+  p = learner$train(task)$predict(task)
+  expect_equal(unname(p$score()), 0)
+})