+++ NEWS.md
@@ -1,3 +1,8 @@
+# mlr3learners 0.1.4
+
+* Fixed a bug in `*.xgboost` and `*.svm` which was triggered if columns
+  were reordered between `$train()` and `$predict()`.
+
 # mlr3learners 0.1.3
 
 * Changes to work with new `mlr3::Learner` API.